logic, markup, styles from the task page separated by components, added components TaskSection and TaskHeader
