+++ src/components/TaskSection/TaskSection.module.scss
@@ -1,0 +1,202 @@
+.current_date_wrapper {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 20px;
+  margin-bottom: 30px;
+  @media screen and (min-width: 768px) {
+    justify-content: start;
+    margin-bottom: 20px;
+  }
+}
+// .pageWrapper {
+//   display: flex;
+//   @media screen and (max-width: 767px) {
+//     flex-direction: column;
+//   }
+// }
+
+.contentWrapper {
+  position: relative;
+  @media screen and (min-width: 768px) and (max-width: 1199px) {
+    padding-left: 20px;
+  }
+  @media screen and (min-width: 1200px) {
+    padding-left: 70px;
+  }
+}
+
+.days_wrapper {
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: flex-end;
+  @media screen and (min-width: 768px) {
+    margin-right: 39px;
+  }
+}
+.back_button {
+  display: inline-block;
+  margin-left: 20px;
+  margin-right: 20px;
+  width: 10px;
+  height: 18px;
+  background-color: transparent;
+  background-image: url('../../img/tasksPage/back.svg');
+  background-repeat: no-repeat;
+  border: none;
+  cursor: pointer;
+  @media screen and (min-width: 768px) {
+    margin-left: 0px;
+  }
+}
+
+.forward_button {
+  display: inline-block;
+  margin-left: 20px;
+  margin-right: 0;
+  width: 10px;
+  height: 18px;
+  background-color: transparent;
+  background-image: url('../../img/tasksPage/forward.svg');
+  background-repeat: no-repeat;
+  border: none;
+  cursor: pointer;
+}
+.сurrent_day {
+  font-family: Montserrat;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 22px;
+  letter-spacing: 0.04em;
+  text-align: left;
+  color: #181c27;
+}
+.slash_days {
+  font-family: Montserrat;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 15px;
+  letter-spacing: 0.04em;
+  text-align: left;
+  color: #181c27;
+}
+
+.duration_days {
+  font-family: Montserrat;
+  font-size: 13px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 16px;
+  letter-spacing: 0.04em;
+  text-align: left;
+  color: #181c27;
+}
+
+.date {
+  font-family: Montserrat;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 15px;
+  letter-spacing: 0.04em;
+  text-align: right;
+  color: rgba(24, 28, 39, 0.6);
+}
+
+.searchFormButtonLabel {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  white-space: nowrap;
+  clip-path: inset(50%);
+  border: 0;
+}
+
+.card_list {
+  display: flex;
+  flex-wrap: wrap;
+  margin-top: 20px;
+  padding-right: 10px;
+  padding-left: 10px;
+
+  @media screen and (min-width: 768px) and (max-width: 1199px) {
+    flex-wrap: wrap;
+    justify-content: space-between;
+  }
+
+  @media screen and (min-width: 1200px) {
+    flex-direction: column;
+    align-items: center;
+  }
+}
+
+.button_analytics {
+  width: 44px;
+  height: 44px;
+  margin-right: 20px;
+  background-image: url(../../img/tasksPage/analytics.png);
+  background-position: 50%;
+  border: none;
+  border-radius: 50%;
+  transition: all 0.3s linear;
+  background-color: #ff6b08;
+  background-repeat: no-repeat;
+  cursor: pointer;
+
+  &:hover {
+    outline: none;
+    box-shadow: 0 0 8px #ff6b08;
+  }
+}
+
+.buttonContainer {
+  position: fixed;
+  width: inherit;
+  display: flex;
+  bottom: 20px;
+  left: 20px;
+  z-index: 20;
+
+  @media screen and (min-width: 767px) {
+    left: 90%;
+  }
+}
+
+// .task_name {
+//   font-family: Montserrat;
+//   font-size: 18px;
+//   font-style: normal;
+//   font-weight: 500;
+//   line-height: 22px;
+//   letter-spacing: 0.04em;
+//   text-align: left;
+// }
+
+// .table_prop {
+//   font-family: Montserrat;
+//   font-size: 10px;
+//   font-style: normal;
+//   font-weight: 400;
+//   line-height: 12px;
+//   letter-spacing: 0.04em;
+//   text-align: left;
+// }
+
+// .table_prop_hour_per_day {
+//   font-family: Montserrat;
+//   font-size: 14px;
+//   font-style: normal;
+//   font-weight: 400;
+//   line-height: 17px;
+//   letter-spacing: 0.04em;
+//   text-align: right;
+// }
+
+// .listItem::before {
+//   background-color: #fff;
+// }
